+++ flaskAPI/app/utils/text_processing.py
@@ -1,0 +1,7 @@
+from konlpy.tag import Okt
+
+okt = Okt()
+
+def extract_nouns(text):
+    nouns = okt.nouns(text)
+    return list(set([n for n in nouns if len(n) > 1]))
